pyproject.toml
[tool.poetry]
name = "details-app"
version = "0.1.0"
description = ""
authors = ["Silent Mobius <alex@vaiolabs.com>"]
readme = "README.md"
packages = [{include = "details_app"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
sqlalchemy = "^2.0.30"
gunicorn = "^22.0.0"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
